requirements.txt
cachetools==4.2.0
DateTime==4.3
exif==1.0.4
future==0.18.2
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
iso8601==0.1.16
more-itertools==8.2.0
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.3
pandas==1.2.0
pefile==2021.9.3
piexif==1.1.3
Pillow==8.1.0
pluggy==0.13.1
plum-py==0.3.1
protobuf==3.14.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
PyExifTool==0.1
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
pytz==2020.5
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rfc3339==6.2
rsa==4.7
Send2Trash==1.8.0
six==1.14.0
spotipy==2.9.0
sqlparse==0.4.1
termcolor==1.1.0
tqdm==4.45.0
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.9
websocket-client==0.57.0
zope.interface==5.2.0
